CategoryDaoへメソッドを移動。 及びActivityServiceの戻り値を修正。
git-svn-id: http://tasklet.googlecode.com/svn/trunk@182 071f029c-86fc-11de-bd5f-d5dc12eeefd8

+++ src/main/java/org/yukung/tasklet/service/impl/ActivityServiceImpl.java
@@ -94,7 +94,7 @@ public class ActivityServiceImpl implements ActivityService {
 	 * @return アクティビティのソート順最大値
 	 */
 	private int getSeq(int userId) {
-		int count = activityDao.getMaxSeqOfActivities(userId);
+		int count = activityDao.getMaxSeqOfActivities(userId).intValue();
 		if (count > 0) {
 			count++;
 		}
